Jan 29, 2026, 4:07 PM

+++ DAA/Assignment_1/binary_search_2d.py
@@ -16,7 +16,7 @@ right = row*col-1
 f = 0
 while left <= right:
     mid = left + (right - left) / 2
-    r = mid / col
+    r = int(mid / col)
     c = mid % col
     cur = arr[r][c]
